requirements.txt
python-telegram-bot~=12.7